// src/types.ts
export interface Job {
  id: number;
  title: string;
  location: string;
  posted_time: string;
  published_at: string;
  job_url: string;
  company_name: string;
  company_url: string;
  description: string;
  applications_count: string;
  contract_type: string;
  experience_level: string;
  work_type: string;
  sector: string;
  salary: string;
  poster_full_name: string;
  poster_profile_url: string;
  company_id: string;
  apply_url: string;
  apply_type: string;
  benefits: string;
}